chore: documentation submodules refactor.
